core: kernel: Switch to Google's GCC 4.9 by default
Legacy GCC 4.8 was dropped in newer builds a long ago. Almost all current
devices use the kernels which were built by GCC 4.9 toolchain. This means
that GCC 4.9 is a more favorable candidate to be a default toolchain. The
devices, whose kernels are not currently buildable via a newer GCC should
set an appropriate TC prefix to legacy one in their trees, though this is
suboptimal.

Change-Id: Ide86ea0d2f559cea08cf93938cbd86429b3bbbf8
Signed-off-by: Alex Saiko <solcmdr@gmail.com>